Revert "added: `linearize!` is now allocation-free"
This reverts commit 25865eacc5486a19b1682eda5ef3923d10174285.
(so `linearize!` is no longer allocation-free)

*MOTIVE*

The intricate solution with a `struct` and 5 separate closures creates a weird type-instability on `NonLinModel` when its `f` and `h` functions are not in the global scope. I'm not sure I fully understand why.

In the end I prefer some small but deterministic allocations in `ForwardDiff.jacobian!` than a type-instability. I also benchmarked the two implementations, and the small allocations are way more performant than the type-instable version.