added: huge performance improvement by caching Jacobians of the constraints
The splatting syntax of `JuMP` forces use to compute the Jacobians of the inequality and equality constraints as a multiple gradients (concatenated in the Jacobian matrices).

This means that `jacobian!` function of the AD tools were called redundantly `ng` and `neq` times, for a specific decision vector value `Z̃`. This is wasteful. A caching mechanism was implemented to store the Jacobians of the constraints and reuse them when needed.

The performance improvement is about 5-10x faster now on `NonLinMPC` with `NonLinModel`.